fix(store): add explicit float8 casts to health score SQL to fix PG18 ambiguity
PostgreSQL 18 rejects `$N * $N` expressions with untyped parameter
placeholders: `ERROR: operator is not unique: unknown * unknown
(SQLSTATE 42725)`. Add explicit `::float8` casts to all parameter
references used in arithmetic within the AtomicRecordHealthResult
query so the multiplication operator resolves unambiguously.

+++ apps/strait/internal/store/health_score.go
@@ -95,35 +95,38 @@ func (q *Queries) AtomicRecordHealthResult(
 
 	// $1=endpoint_url $2=successVal $3=timeoutVal $4=latencyVal
 	// $5=alpha $6=wSuccess $7=wTimeout $8=wLatency $9=lastLatencyMs
+	//
+	// All parameter references are explicitly cast to float8 to avoid
+	// "operator is not unique: unknown * unknown" on PostgreSQL 18+.
 	const atomicSQL = `
 		INSERT INTO endpoint_health_scores AS ehs (
 			endpoint_url, success_rate, timeout_rate, latency_score,
 			health_score, total_requests, last_latency_ms, updated_at
 		) VALUES (
 			$1,
-			$5 * $2 + (1.0 - $5) * 1.0,
-			$5 * $3 + (1.0 - $5) * 0.0,
-			$5 * $4 + (1.0 - $5) * 1.0,
+			$5::float8 * $2::float8 + (1.0 - $5::float8) * 1.0,
+			$5::float8 * $3::float8 + (1.0 - $5::float8) * 0.0,
+			$5::float8 * $4::float8 + (1.0 - $5::float8) * 1.0,
 			LEAST(100.0, GREATEST(0.0,
-				($6 * ($5 * $2 + (1.0 - $5) * 1.0) +
-				 $7 * (1.0 - ($5 * $3 + (1.0 - $5) * 0.0)) +
-				 $8 * ($5 * $4 + (1.0 - $5) * 1.0)) * 100.0
+				($6::float8 * ($5::float8 * $2::float8 + (1.0 - $5::float8) * 1.0) +
+				 $7::float8 * (1.0 - ($5::float8 * $3::float8 + (1.0 - $5::float8) * 0.0)) +
+				 $8::float8 * ($5::float8 * $4::float8 + (1.0 - $5::float8) * 1.0)) * 100.0
 			)),
 			1,
-			$9,
+			$9::float8,
 			NOW()
 		)
 		ON CONFLICT (endpoint_url) DO UPDATE SET
-			success_rate  = $5 * $2 + (1.0 - $5) * ehs.success_rate,
-			timeout_rate  = $5 * $3 + (1.0 - $5) * ehs.timeout_rate,
-			latency_score = $5 * $4 + (1.0 - $5) * ehs.latency_score,
+			success_rate  = $5::float8 * $2::float8 + (1.0 - $5::float8) * ehs.success_rate,
+			timeout_rate  = $5::float8 * $3::float8 + (1.0 - $5::float8) * ehs.timeout_rate,
+			latency_score = $5::float8 * $4::float8 + (1.0 - $5::float8) * ehs.latency_score,
 			health_score  = LEAST(100.0, GREATEST(0.0,
-				($6 * ($5 * $2 + (1.0 - $5) * ehs.success_rate) +
-				 $7 * (1.0 - ($5 * $3 + (1.0 - $5) * ehs.timeout_rate)) +
-				 $8 * ($5 * $4 + (1.0 - $5) * ehs.latency_score)) * 100.0
+				($6::float8 * ($5::float8 * $2::float8 + (1.0 - $5::float8) * ehs.success_rate) +
+				 $7::float8 * (1.0 - ($5::float8 * $3::float8 + (1.0 - $5::float8) * ehs.timeout_rate)) +
+				 $8::float8 * ($5::float8 * $4::float8 + (1.0 - $5::float8) * ehs.latency_score)) * 100.0
 			)),
 			total_requests = ehs.total_requests + 1,
-			last_latency_ms = $9,
+			last_latency_ms = $9::float8,
 			updated_at = NOW()
 		RETURNING endpoint_url, health_score, success_rate, timeout_rate, latency_score,
 		          total_requests, last_latency_ms, updated_at, created_at
